+++ src/flow.ts
@@ -1,18 +1,18 @@
-export { Chord, IChordSettings, IChordDataItem } from "./.internal/charts/flow/Chord";
-export { ChordDirected, IChordDirectedSettings, IChordDirectedDataItem } from "./.internal/charts/flow/ChordDirected";
-export { ChordNonRibbon, IChordNonRibbonSettings, IChordNonRibbonDataItem } from "./.internal/charts/flow/ChordNonRibbon";
-export { ChordLink, IChordLinkSettings } from "./.internal/charts/flow/ChordLink";
-export { ChordLinkDirected, IChordLinkDirectedSettings } from "./.internal/charts/flow/ChordLinkDirected";
-export { ChordNodes, IChordNodesSettings, IChordNodesDataItem } from "./.internal/charts/flow/ChordNodes";
+export { Chord, IChordSettings, IChordDataItem, IChordEvents, IChordPrivate } from "./.internal/charts/flow/Chord";
+export { ChordDirected, IChordDirectedSettings, IChordDirectedDataItem, IChordDirectedEvents, IChordDirectedPrivate } from "./.internal/charts/flow/ChordDirected";
+export { ChordNonRibbon, IChordNonRibbonSettings, IChordNonRibbonDataItem, IChordNonRibbonEvents, IChordNonRibbonPrivate } from "./.internal/charts/flow/ChordNonRibbon";
+export { ChordLink, IChordLinkSettings, IChordLinkPrivate } from "./.internal/charts/flow/ChordLink";
+export { ChordLinkDirected, IChordLinkDirectedSettings, IChordLinkDirectedPrivate } from "./.internal/charts/flow/ChordLinkDirected";
+export { ChordNodes, IChordNodesSettings, IChordNodesDataItem, IChordNodesEvents, IChordNodesPrivate } from "./.internal/charts/flow/ChordNodes";
 
-export { Flow, IFlowSettings, IFlowDataItem } from "./.internal/charts/flow/Flow";
-export { FlowLink, IFlowLinkSettings } from "./.internal/charts/flow/FlowLink";
-export { FlowNode, IFlowNodeSettings } from "./.internal/charts/flow/FlowNode";
-export { FlowNodes, IFlowNodesSettings, IFlowNodesDataItem } from "./.internal/charts/flow/FlowNodes";
+export { Flow, IFlowSettings, IFlowDataItem, IFlowEvents, IFlowPrivate } from "./.internal/charts/flow/Flow";
+export { FlowLink, IFlowLinkSettings, IFlowLinkPrivate } from "./.internal/charts/flow/FlowLink";
+export { FlowNode, IFlowNodeSettings, IFlowNodePrivate } from "./.internal/charts/flow/FlowNode";
+export { FlowNodes, IFlowNodesSettings, IFlowNodesDataItem, IFlowNodesEvents, IFlowNodesPrivate } from "./.internal/charts/flow/FlowNodes";
 
-export { Sankey, ISankeySettings, ISankeyDataItem } from "./.internal/charts/flow/Sankey";
-export { SankeyLink, ISankeyLinkSettings } from "./.internal/charts/flow/SankeyLink";
-export { SankeyNodes, ISankeyNodesSettings, ISankeyNodesDataItem } from "./.internal/charts/flow/SankeyNodes";
+export { Sankey, ISankeySettings, ISankeyDataItem, ISankeyEvents, ISankeyPrivate } from "./.internal/charts/flow/Sankey";
+export { SankeyLink, ISankeyLinkSettings, ISankeyLinkPrivate } from "./.internal/charts/flow/SankeyLink";
+export { SankeyNodes, ISankeyNodesSettings, ISankeyNodesDataItem, ISankeyNodesEvents, ISankeyNodesPrivate } from "./.internal/charts/flow/SankeyNodes";
 
 import { FlowDefaultTheme } from "./.internal/charts/flow/FlowDefaultTheme";
 export { FlowDefaultTheme as DefaultTheme };