+++ src/hierarchy.ts
@@ -1,18 +1,18 @@
-export { BreadcrumbBar, IBreadcrumbBarSettings } from "./.internal/charts/hierarchy/BreadcrumbBar";
-export { ForceDirected, IForceDirectedSettings, IForceDirectedDataItem } from "./.internal/charts/hierarchy/ForceDirected";
+export { BreadcrumbBar, IBreadcrumbBarSettings, IBreadcrumbBarEvents, IBreadcrumbBarPrivate } from "./.internal/charts/hierarchy/BreadcrumbBar";
+export { ForceDirected, IForceDirectedSettings, IForceDirectedDataItem, IForceDirectedDataObject, IForceDirectedEvents, IForceDirectedPrivate } from "./.internal/charts/hierarchy/ForceDirected";
 
-export { Hierarchy, IHierarchySettings, IHierarchyDataItem } from "./.internal/charts/hierarchy/Hierarchy";
-export { HierarchyLink, IHierarchyLinkSettings } from "./.internal/charts/hierarchy/HierarchyLink";
-export { HierarchyNode, IHierarchyNodeSettings } from "./.internal/charts/hierarchy/HierarchyNode";
+export { Hierarchy, IHierarchySettings, IHierarchyDataItem, IHierarchyDataObject, IHierarchyEvents, IHierarchyPrivate } from "./.internal/charts/hierarchy/Hierarchy";
+export { HierarchyLink, IHierarchyLinkSettings, IHierarchyLinkPrivate } from "./.internal/charts/hierarchy/HierarchyLink";
+export { HierarchyNode, IHierarchyNodeSettings, IHierarchyNodePrivate } from "./.internal/charts/hierarchy/HierarchyNode";
 
-export { LinkedHierarchy, ILinkedHierarchySettings, ILinkedHierarchyDataItem } from "./.internal/charts/hierarchy/LinkedHierarchy";
-export { LinkedHierarchyNode, ILinkedHierarchyNodeSettings } from "./.internal/charts/hierarchy/LinkedHierarchyNode";
+export { LinkedHierarchy, ILinkedHierarchySettings, ILinkedHierarchyDataItem, ILinkedHierarchyDataObject, ILinkedHierarchyEvents, ILinkedHierarchyPrivate } from "./.internal/charts/hierarchy/LinkedHierarchy";
+export { LinkedHierarchyNode, ILinkedHierarchyNodeSettings, ILinkedHierarchyNodePrivate } from "./.internal/charts/hierarchy/LinkedHierarchyNode";
 
-export { Pack, IPackSettings, IPackDataItem } from "./.internal/charts/hierarchy/Pack";
-export { Partition, IPartitionSettings, IPartitionDataItem } from "./.internal/charts/hierarchy/Partition";
-export { Sunburst, ISunburstSettings, ISunburstDataItem } from "./.internal/charts/hierarchy/Sunburst";
-export { Tree, ITreeSettings, ITreeDataItem } from "./.internal/charts/hierarchy/Tree";
-export { Treemap, ITreemapSettings, ITreemapDataItem } from "./.internal/charts/hierarchy/Treemap";
+export { Pack, IPackSettings, IPackDataItem, IPackDataObject, IPackEvents, IPackPrivate } from "./.internal/charts/hierarchy/Pack";
+export { Partition, IPartitionSettings, IPartitionDataItem, IPartitionDataObject, IPartitionPrivate } from "./.internal/charts/hierarchy/Partition";
+export { Sunburst, ISunburstSettings, ISunburstDataItem, ISunburstDataObject, ISunburstPrivate } from "./.internal/charts/hierarchy/Sunburst";
+export { Tree, ITreeSettings, ITreeDataItem, ITreeDataObject, ITreeEvents, ITreePrivate } from "./.internal/charts/hierarchy/Tree";
+export { Treemap, ITreemapSettings, ITreemapDataItem, ITreemapDataObject, ITreemapPrivate } from "./.internal/charts/hierarchy/Treemap";
 
 import { HierarchyDefaultTheme } from "./.internal/charts/hierarchy/HierarchyDefaultTheme";
 export { HierarchyDefaultTheme as DefaultTheme };