+++ src/map.ts
@@ -1,12 +1,12 @@
-export { GraticuleSeries, IGraticuleSeriesSettings, IGraticuleSeriesDataItem } from "./.internal/charts/map/GraticuleSeries";
-export { MapChart, IMapChartSettings } from "./.internal/charts/map/MapChart";
-export { MapLine, IMapLineSettings } from "./.internal/charts/map/MapLine";
-export { MapLineSeries, IMapLineSeriesSettings, IMapLineSeriesDataItem } from "./.internal/charts/map/MapLineSeries";
-export { MapPointSeries, IMapPointSeriesSettings } from "./.internal/charts/map/MapPointSeries";
-export { MapPolygon, IMapPolygonSettings } from "./.internal/charts/map/MapPolygon";
-export { MapPolygonSeries, IMapPolygonSeriesSettings, IMapPolygonSeriesDataItem } from "./.internal/charts/map/MapPolygonSeries";
-export { MapSeries, IMapSeriesSettings, IMapSeriesDataItem } from "./.internal/charts/map/MapSeries";
-export { ZoomControl, IZoomControlSettings } from "./.internal/charts/map/ZoomControl";
+export { GraticuleSeries, IGraticuleSeriesSettings, IGraticuleSeriesDataItem, IGraticuleSeriesPrivate } from "./.internal/charts/map/GraticuleSeries";
+export { MapChart, IMapChartSettings, IMapChartEvents, IMapChartPrivate } from "./.internal/charts/map/MapChart";
+export { MapLine, IMapLineSettings, IMapLinePrivate } from "./.internal/charts/map/MapLine";
+export { MapLineSeries, IMapLineSeriesSettings, IMapLineSeriesDataItem, IMapLineSeriesPrivate } from "./.internal/charts/map/MapLineSeries";
+export { MapPointSeries, IMapPointSeriesSettings, IMapPointSeriesDataItem, IMapPointSeriesPrivate } from "./.internal/charts/map/MapPointSeries";
+export { MapPolygon, IMapPolygonSettings, IMapPolygonPrivate } from "./.internal/charts/map/MapPolygon";
+export { MapPolygonSeries, IMapPolygonSeriesSettings, IMapPolygonSeriesDataItem, IMapPolygonSeriesPrivate } from "./.internal/charts/map/MapPolygonSeries";
+export { MapSeries, IMapSeriesSettings, IMapSeriesDataItem, IMapSeriesEvents, IMapSeriesPrivate } from "./.internal/charts/map/MapSeries";
+export { ZoomControl, IZoomControlSettings, IZoomControlPrivate } from "./.internal/charts/map/ZoomControl";
 export { getGeoRectangle, getGeoCircle, normalizeGeoPoint } from "./.internal/charts/map/MapUtils"
 
 import { geoMercator, geoOrthographic, geoEquirectangular, geoAlbersUsa, geoEqualEarth, geoNaturalEarth1 } from "d3-geo";