+++ src/percent.ts
@@ -1,14 +1,14 @@
-export { PercentChart, IPercentChartSettings } from "./.internal/charts/percent/PercentChart";
-export { PercentSeries, IPercentSeriesSettings, IPercentSeriesDataItem } from "./.internal/charts/percent/PercentSeries";
+export { PercentChart, IPercentChartSettings, IPercentChartPrivate } from "./.internal/charts/percent/PercentChart";
+export { PercentSeries, IPercentSeriesSettings, IPercentSeriesDataItem, IPercentSeriesPrivate } from "./.internal/charts/percent/PercentSeries";
 
-export { PieChart, IPieChartSettings } from "./.internal/charts/pie/PieChart";
-export { PieSeries, IPieSeriesSettings, IPieSeriesDataItem } from "./.internal/charts/pie/PieSeries";
+export { PieChart, IPieChartSettings, IPieChartPrivate } from "./.internal/charts/pie/PieChart";
+export { PieSeries, IPieSeriesSettings, IPieSeriesDataItem, IPieSeriesPrivate } from "./.internal/charts/pie/PieSeries";
 
-export { FunnelSeries, IFunnelSeriesSettings, IFunnelSeriesDataItem } from "./.internal/charts/funnel/FunnelSeries";
-export { FunnelSlice } from "./.internal/charts/funnel/FunnelSlice";
-export { PictorialStackedSeries, IPictorialStackedSeriesSettings, IPictorialStackedSeriesDataItem } from "./.internal/charts/funnel/PictorialStackedSeries";
-export { PyramidSeries, IPyramidSeriesSettings, IPyramidSeriesDataItem } from "./.internal/charts/funnel/PyramidSeries";
-export { SlicedChart, ISlicedChartSettings } from "./.internal/charts/funnel/SlicedChart";
+export { FunnelSeries, IFunnelSeriesSettings, IFunnelSeriesDataItem, IFunnelSeriesPrivate } from "./.internal/charts/funnel/FunnelSeries";
+export { FunnelSlice, IFunnelSlicePrivate, IFunnelSliceSettings } from "./.internal/charts/funnel/FunnelSlice";
+export { PictorialStackedSeries, IPictorialStackedSeriesSettings, IPictorialStackedSeriesDataItem, IPictorialStackedSeriesPrivate } from "./.internal/charts/funnel/PictorialStackedSeries";
+export { PyramidSeries, IPyramidSeriesSettings, IPyramidSeriesDataItem, IPyramidSeriesPrivate } from "./.internal/charts/funnel/PyramidSeries";
+export { SlicedChart, ISlicedChartSettings, ISlicedChartPrivate } from "./.internal/charts/funnel/SlicedChart";
 
 import { PercentDefaultTheme } from "./.internal/charts/percent/PercentDefaultTheme";
 export { PercentDefaultTheme as DefaultTheme };