+++ src/radar.ts
@@ -1,11 +1,11 @@
-export { AxisRendererCircular, IAxisRendererCircularSettings } from "./.internal/charts/radar/AxisRendererCircular";
-export { AxisRendererRadial, IAxisRendererRadialSettings } from "./.internal/charts/radar/AxisRendererRadial";
-export { ClockHand, IClockHandSettings } from "./.internal/charts/radar/ClockHand";
-export { RadarChart, IRadarChartSettings } from "./.internal/charts/radar/RadarChart";
-export { RadarColumnSeries, IRadarColumnSeriesSettings, IRadarColumnSeriesDataItem } from "./.internal/charts/radar/RadarColumnSeries";
-export { RadarCursor, IRadarCursorSettings } from "./.internal/charts/radar/RadarCursor";
-export { RadarLineSeries, IRadarLineSeriesSettings, IRadarLineSeriesDataItem } from "./.internal/charts/radar/RadarLineSeries";
-export { SmoothedRadarLineSeries, ISmoothedRadarLineSeriesSettings, ISmoothedRadarLineSeriesDataItem } from "./.internal/charts/radar/SmoothedRadarLineSeries";
+export { AxisRendererCircular, IAxisRendererCircularSettings, IAxisRendererCircularPrivate } from "./.internal/charts/radar/AxisRendererCircular";
+export { AxisRendererRadial, IAxisRendererRadialSettings, IAxisRendererRadialPrivate } from "./.internal/charts/radar/AxisRendererRadial";
+export { ClockHand, IClockHandSettings, IClockHandPrivate } from "./.internal/charts/radar/ClockHand";
+export { RadarChart, IRadarChartSettings, IRadarChartPrivate } from "./.internal/charts/radar/RadarChart";
+export { RadarColumnSeries, IRadarColumnSeriesSettings, IRadarColumnSeriesDataItem, IRadarColumnSeriesAxisRange, IRadarColumnSeriesPrivate } from "./.internal/charts/radar/RadarColumnSeries";
+export { RadarCursor, IRadarCursorSettings, IRadarCursorEvents, IRadarCursorPrivate } from "./.internal/charts/radar/RadarCursor";
+export { RadarLineSeries, IRadarLineSeriesSettings, IRadarLineSeriesDataItem, IRadarLineSeriesAxisRange, IRadarLineSeriesPrivate } from "./.internal/charts/radar/RadarLineSeries";
+export { SmoothedRadarLineSeries, ISmoothedRadarLineSeriesSettings, ISmoothedRadarLineSeriesDataItem, ISmoothedRadarLineSeriesPrivate } from "./.internal/charts/radar/SmoothedRadarLineSeries";
 
 import { RadarDefaultTheme } from "./.internal/charts/radar/RadarDefaultTheme";
 export { RadarDefaultTheme as DefaultTheme };