+++ src/venn.ts
@@ -1,3 +1,3 @@
-export { Venn, IVennSettings } from "./.internal/charts/venn/Venn";
+export { Venn, IVennSettings, IVennDataItem, IVennPrivate } from "./.internal/charts/venn/Venn";
 import { VennDefaultTheme } from "./.internal/charts/venn/VennDefaultTheme";
 export { VennDefaultTheme as DefaultTheme };