+++ src/wc.ts
@@ -1,3 +1,3 @@
-export { WordCloud, IWordCloudSettings } from "./.internal/charts/wordcloud/WordCloud";
+export { WordCloud, IWordCloudSettings, IWordCloudDataItem, IWordCloudPrivate } from "./.internal/charts/wordcloud/WordCloud";
 import { WordCloudDefaultTheme } from "./.internal/charts/wordcloud/WordCloudDefaultTheme";
 export { WordCloudDefaultTheme as DefaultTheme };