+++ src/xy.ts
@@ -1,40 +1,40 @@
-export { XYChart, IXYChartSettings } from "./.internal/charts/xy/XYChart";
-export { XYChartScrollbar, IXYChartScrollbarSettings } from "./.internal/charts/xy/XYChartScrollbar";
-export { XYCursor, IXYCursorSettings } from "./.internal/charts/xy/XYCursor";
-export { BaseColumnSeries } from "./.internal/charts/xy/series/BaseColumnSeries";
+export { XYChart, IXYChartSettings, IXYChartEvents, IXYChartPrivate } from "./.internal/charts/xy/XYChart";
+export { XYChartScrollbar, IXYChartScrollbarSettings, IXYChartScrollbarPrivate } from "./.internal/charts/xy/XYChartScrollbar";
+export { XYCursor, IXYCursorSettings, IXYCursorEvents, IXYCursorPrivate } from "./.internal/charts/xy/XYCursor";
+export { BaseColumnSeries, IBaseColumnSeriesAxisRange, IBaseColumnSeriesDataItem, IBaseColumnSeriesPrivate, IBaseColumnSeriesSettings } from "./.internal/charts/xy/series/BaseColumnSeries";
 
-export { Axis, IAxisSettings, IAxisDataItem } from "./.internal/charts/xy/axes/Axis";
-export { CategoryAxis, ICategoryAxisSettings, ICategoryAxisDataItem } from "./.internal/charts/xy/axes/CategoryAxis";
-export { CategoryDateAxis, ICategoryDateAxisSettings } from "./.internal/charts/xy/axes/CategoryDateAxis";
-export { DateAxis, IDateAxisSettings, IDateAxisDataItem } from "./.internal/charts/xy/axes/DateAxis";
-export { GaplessDateAxis, IGaplessDateAxisSettings } from "./.internal/charts/xy/axes/GaplessDateAxis";
-export { DurationAxis, IDurationAxisSettings, IDurationAxisDataItem } from "./.internal/charts/xy/axes/DurationAxis";
-export { ValueAxis, IValueAxisSettings, IValueAxisDataItem } from "./.internal/charts/xy/axes/ValueAxis";
+export { Axis, IAxisSettings, IAxisDataItem, IAxisEvents, IAxisPrivate } from "./.internal/charts/xy/axes/Axis";
+export { CategoryAxis, ICategoryAxisSettings, ICategoryAxisDataItem, ICategoryAxisPrivate, ICategoryAxisEvents } from "./.internal/charts/xy/axes/CategoryAxis";
+export { CategoryDateAxis, ICategoryDateAxisSettings, ICategoryDateAxisDataItem, ICategoryDateAxisEvents, ICategoryDateAxisPrivate } from "./.internal/charts/xy/axes/CategoryDateAxis";
+export { DateAxis, IDateAxisSettings, IDateAxisDataItem, IDateAxisEvents, IDateAxisPrivate } from "./.internal/charts/xy/axes/DateAxis";
+export { GaplessDateAxis, IGaplessDateAxisSettings, IGaplessDateAxisDataItem, IGaplessDateAxisEvents, IGaplessDateAxisPrivate } from "./.internal/charts/xy/axes/GaplessDateAxis";
+export { DurationAxis, IDurationAxisSettings, IDurationAxisDataItem, IDurationAxisEvents, IDurationAxisPrivate } from "./.internal/charts/xy/axes/DurationAxis";
+export { ValueAxis, IValueAxisSettings, IValueAxisDataItem, IMinMaxStep, IValueAxisEvents, IValueAxisPrivate } from "./.internal/charts/xy/axes/ValueAxis";
 
-export { AxisBullet, IAxisBulletSettings } from "./.internal/charts/xy/axes/AxisBullet";
-export { AxisLabel, IAxisLabelSettings } from "./.internal/charts/xy/axes/AxisLabel";
-export { AxisLabelRadial, IAxisLabelRadialSettings } from "./.internal/charts/xy/axes/AxisLabelRadial";
-export { AxisTick, IAxisTickSettings } from "./.internal/charts/xy/axes/AxisTick";
-export { Grid, IGridSettings } from "./.internal/charts/xy/axes/Grid";
+export { AxisBullet, IAxisBulletSettings, IAxisBulletPrivate } from "./.internal/charts/xy/axes/AxisBullet";
+export { AxisLabel, IAxisLabelSettings, IAxisLabel, IAxisLabelPrivate } from "./.internal/charts/xy/axes/AxisLabel";
+export { AxisLabelRadial, IAxisLabelRadialSettings, IAxisLabelRadialPrivate } from "./.internal/charts/xy/axes/AxisLabelRadial";
+export { AxisTick, IAxisTickSettings, IAxisTickPrivate } from "./.internal/charts/xy/axes/AxisTick";
+export { Grid, IGridSettings, IGridPrivate } from "./.internal/charts/xy/axes/Grid";
 
-export { AxisRenderer, IAxisRendererSettings } from "./.internal/charts/xy/axes/AxisRenderer";
-export { AxisRendererX, IAxisRendererXSettings } from "./.internal/charts/xy/axes/AxisRendererX";
-export { AxisRendererY, IAxisRendererYSettings } from "./.internal/charts/xy/axes/AxisRendererY";
+export { AxisRenderer, IAxisRendererSettings, IAxisRendererPrivate } from "./.internal/charts/xy/axes/AxisRenderer";
+export { AxisRendererX, IAxisRendererXSettings, IAxisRendererXPrivate } from "./.internal/charts/xy/axes/AxisRendererX";
+export { AxisRendererY, IAxisRendererYSettings, IAxisRendererYPrivate } from "./.internal/charts/xy/axes/AxisRendererY";
 
-export { XYSeries, IXYSeriesSettings, IXYSeriesDataItem } from "./.internal/charts/xy/series/XYSeries";
+export { XYSeries, IXYSeriesSettings, IXYSeriesDataItem, IXYSeriesAxisRange, IXYSeriesEvents, IXYSeriesPrivate } from "./.internal/charts/xy/series/XYSeries";
 
-export { Candlestick, ICandlestickSettings } from "./.internal/charts/xy/series/Candlestick";
-export { CandlestickSeries, ICandlestickSeriesSettings, ICandlestickSeriesDataItem } from "./.internal/charts/xy/series/CandlestickSeries";
+export { Candlestick, ICandlestickSettings, ICandlestickPrivate } from "./.internal/charts/xy/series/Candlestick";
+export { CandlestickSeries, ICandlestickSeriesSettings, ICandlestickSeriesDataItem, ICandlestickSeriesAxisRange, ICandlestickSeriesPrivate } from "./.internal/charts/xy/series/CandlestickSeries";
 
-export { OHLC, IOHLCSettings } from "./.internal/charts/xy/series/OHLC";
-export { OHLCSeries, IOHLCSeriesSettings, IOHLCSeriesDataItem } from "./.internal/charts/xy/series/OHLCSeries";
+export { OHLC, IOHLCSettings, IOHLCPrivate } from "./.internal/charts/xy/series/OHLC";
+export { OHLCSeries, IOHLCSeriesSettings, IOHLCSeriesDataItem, IOHLCSeriesAxisRange, IOHLCSeriesPrivate } from "./.internal/charts/xy/series/OHLCSeries";
 
-export { ColumnSeries, IColumnSeriesSettings, IColumnSeriesDataItem } from "./.internal/charts/xy/series/ColumnSeries";
-export { LineSeries, ILineSeriesSettings, ILineSeriesDataItem } from "./.internal/charts/xy/series/LineSeries";
-export { SmoothedYLineSeries, ISmoothedYLineSeriesSettings, ISmoothedYLineSeriesDataItem } from "./.internal/charts/xy/series/SmoothedYLineSeries";
-export { SmoothedXLineSeries, ISmoothedXLineSeriesSettings, ISmoothedXLineSeriesDataItem } from "./.internal/charts/xy/series/SmoothedXLineSeries";
-export { SmoothedXYLineSeries, ISmoothedXYLineSeriesDataItem } from "./.internal/charts/xy/series/SmoothedXYLineSeries";
-export { StepLineSeries, IStepLineSeriesSettings, IStepLineSeriesDataItem } from "./.internal/charts/xy/series/StepLineSeries";
+export { ColumnSeries, IColumnSeriesSettings, IColumnSeriesDataItem, IColumnSeriesAxisRange, IColumnSeriesPrivate } from "./.internal/charts/xy/series/ColumnSeries";
+export { LineSeries, ILineSeriesSettings, ILineSeriesDataItem, ILineSeriesAxisRange, ILineSeriesPrivate } from "./.internal/charts/xy/series/LineSeries";
+export { SmoothedYLineSeries, ISmoothedYLineSeriesSettings, ISmoothedYLineSeriesDataItem, ISmoothedYLineSeriesPrivate } from "./.internal/charts/xy/series/SmoothedYLineSeries";
+export { SmoothedXLineSeries, ISmoothedXLineSeriesSettings, ISmoothedXLineSeriesDataItem, ISmoothedXLineSeriesPrivate } from "./.internal/charts/xy/series/SmoothedXLineSeries";
+export { SmoothedXYLineSeries, ISmoothedXYLineSeriesDataItem, SmoothedXYLineSeriesDataItem, SmoothedXYLineSeriesPrivate, SmoothedXYLineSeriesProperties } from "./.internal/charts/xy/series/SmoothedXYLineSeries";
+export { StepLineSeries, IStepLineSeriesSettings, IStepLineSeriesDataItem, IStepLineSeriesPrivate } from "./.internal/charts/xy/series/StepLineSeries";
 
 import { XYChartDefaultTheme } from "./.internal/charts/xy/XYChartDefaultTheme";
 export { XYChartDefaultTheme as DefaultTheme };